Implemented feedback after adding replies
